refactor: HealthCheckResponse 패키지 이동
