+++ app-scheduler/src/main/java/com/parkingcomestrue/external/api/HealthChecker.java
@@ -1,7 +1,5 @@
 package com.parkingcomestrue.external.api;
 
-import com.parkingcomestrue.external.api.parkingapi.HealthCheckResponse;
-
 public interface HealthChecker {
 
     HealthCheckResponse check();
